Tidy up Contacts logic
Moves unrelated functionality out of the Contacts object, to simplify
the layout. This will allow us to move things about more easily in the
future.
