+++ anghammarad/src/main/scala/com/gu/anghammarad/Enrichments.scala
@@ -1,8 +1,7 @@
 package com.gu.anghammarad
 
-import cats.syntax.either._
+import scala.util.{Success, Try}
 
-import scala.util.{Failure, Success, Try}
 
 object Enrichments {
   implicit class RichList[A](as: List[A]) {
@@ -22,13 +21,6 @@ object Enrichments {
           bs <- acc
         } yield b :: bs
       }
-    }
-  }
-
-  implicit class RichTry[A](underlying: Try[A]) {
-    def toEither: Either[Throwable, A] = underlying match {
-      case Failure(e) => Left(e)
-      case Success(a) => Right(a)
     }
   }
 }
